batch: Extract absence content builder
Batch ownership stores absence claims with content held in LineBuffer
instances.

The streaming builder for that content sits inside batch.ownership, so
callers that need absence-content construction import the wider ownership
model.

This commit moves absence-content construction into batch.absence_content
and updates ownership plus sift result planning to use the new owner.

Subsequent commits will move tests and architecture guards to the new
module boundary.
